requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
blinker==1.8.2
    # via flask
certifi==2024.6.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via flask
colorama==0.4.6
    # via click
ecdsa==0.19.0
    # via python-jose
flask==3.0.3
    # via -r requirements.in
idna==3.7
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
pyasn1==0.6.0
    # via
    #   python-jose
    #   rsa
python-jose==3.3.0
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
rsa==4.9
    # via python-jose
six==1.16.0
    # via ecdsa
urllib3==2.2.2
    # via requests
werkzeug==3.0.3
    # via flask
